refactor: Update exception's parent
