+++ src/PhpPact/Consumer/Driver/Body/InteractionBodyDriver.php
@@ -5,8 +5,8 @@ namespace PhpPact\Consumer\Driver\Body;
 use FFI;
 use FFI\CData;
 use PhpPact\Consumer\Driver\Enum\InteractionPart;
-use PhpPact\Consumer\Exception\InteractionBodyNotAddedException;
-use PhpPact\Consumer\Exception\PartNotAddedException;
+use PhpPact\Consumer\Driver\Exception\InteractionBodyNotAddedException;
+use PhpPact\Consumer\Driver\Exception\PartNotAddedException;
 use PhpPact\Consumer\Model\Body\Binary;
 use PhpPact\Consumer\Model\Body\Multipart;
 use PhpPact\Consumer\Model\Body\Text;
