+++ src/PhpPact/Consumer/Driver/Exception/InteractionBodyNotAddedException.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace PhpPact\Consumer\Driver\Exception;
+
+class InteractionBodyNotAddedException extends DriverException
+{
+}
