+++ src/PhpPact/Consumer/Driver/Exception/MessageContentsNotAddedException.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace PhpPact\Consumer\Driver\Exception;
+
+class MessageContentsNotAddedException extends DriverException
+{
+}
