+++ src/PhpPact/Consumer/Driver/Exception/PactFileNotWroteException.php
@@ -1,10 +1,8 @@
 <?php
 
-namespace PhpPact\Consumer\Exception;
+namespace PhpPact\Consumer\Driver\Exception;
 
-use Exception;
-
-class PactFileNotWroteException extends Exception
+class PactFileNotWroteException extends DriverException
 {
     public function __construct(int $code)
     {
