+++ src/PhpPact/Consumer/Driver/Exception/PartNotAddedException.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace PhpPact\Consumer\Driver\Exception;
+
+class PartNotAddedException extends DriverException
+{
+}
