+++ src/PhpPact/Consumer/Exception/BinaryFileNotExistException.php
@@ -2,8 +2,6 @@
 
 namespace PhpPact\Consumer\Exception;
 
-use Exception;
-
-class BinaryFileNotExistException extends Exception
+class BinaryFileNotExistException extends ConsumerException
 {
 }
