+++ src/PhpPact/Consumer/Exception/BodyNotSupportedException.php
@@ -2,8 +2,6 @@
 
 namespace PhpPact\Consumer\Exception;
 
-use Exception;
-
-class BodyNotSupportedException extends Exception
+class BodyNotSupportedException extends ConsumerException
 {
 }
