--- src/PhpPact/Consumer/Exception/InteractionBodyNotAddedException.php
@@ -1,9 +1,0 @@
-<?php
-
-namespace PhpPact\Consumer\Exception;
-
-use Exception;
-
-class InteractionBodyNotAddedException extends Exception
-{
-}
