--- src/PhpPact/Consumer/Exception/MessageContentsNotAddedException.php
@@ -1,9 +1,0 @@
-<?php
-
-namespace PhpPact\Consumer\Exception;
-
-use Exception;
-
-class MessageContentsNotAddedException extends Exception
-{
-}
