--- src/PhpPact/Consumer/Exception/PartNotAddedException.php
@@ -1,9 +1,0 @@
-<?php
-
-namespace PhpPact\Consumer\Exception;
-
-use Exception;
-
-class PartNotAddedException extends Exception
-{
-}
