+++ src/PhpPact/Plugin/Exception/PluginException.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace PhpPact\Plugin\Exception;
+
+use PhpPact\Exception\BaseException;
+
+class PluginException extends BaseException
+{
+}
