+++ src/PhpPact/Standalone/StubService/Exception/StubServerException.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace PhpPact\Standalone\StubService\Exception;
+
+use PhpPact\Exception\BaseException;
+
+class StubServerException extends BaseException
+{
+}
