+++ tests/PhpPact/Consumer/Driver/Body/InteractionBodyDriverTest.php
@@ -7,8 +7,8 @@ use FFI\CData;
 use PhpPact\Consumer\Driver\Body\InteractionBodyDriver;
 use PhpPact\Consumer\Driver\Body\InteractionBodyDriverInterface;
 use PhpPact\Consumer\Driver\Enum\InteractionPart;
-use PhpPact\Consumer\Exception\InteractionBodyNotAddedException;
-use PhpPact\Consumer\Exception\PartNotAddedException;
+use PhpPact\Consumer\Driver\Exception\InteractionBodyNotAddedException;
+use PhpPact\Consumer\Driver\Exception\PartNotAddedException;
 use PhpPact\Consumer\Model\Body\Binary;
 use PhpPact\Consumer\Model\Body\Multipart;
 use PhpPact\Consumer\Model\Body\Part;
