+++ tests/PhpPact/Consumer/Driver/Body/MessageBodyDriverTest.php
@@ -4,7 +4,7 @@ namespace PhpPactTest\Consumer\Driver\Body;
 
 use PhpPact\Consumer\Driver\Body\MessageBodyDriver;
 use PhpPact\Consumer\Driver\Body\MessageBodyDriverInterface;
-use PhpPact\Consumer\Exception\MessageContentsNotAddedException;
+use PhpPact\Consumer\Driver\Exception\MessageContentsNotAddedException;
 use PhpPact\Consumer\Model\Body\Binary;
 use PhpPact\Consumer\Model\Body\Text;
 use PhpPact\Consumer\Model\Message;
